+++ src/data/types/nucleotides/bench.rs
@@ -22,7 +22,7 @@ fn translate_sequence_long(b: &mut Bencher) {
 fn validate_retain_iupac_uc(b: &mut Bencher) {
     b.iter(|| {
         SEQ.clone().retain_mut(|b| {
-            *b = TO_DNA_IUPAC_NO_GAPS_UC[*b as usize];
+            *b = TO_DNA_IUPAC_NO_GAPS_UC[*b];
             *b > 0
         });
     });
@@ -35,7 +35,7 @@ fn validate_filtermap_iupac_uc(b: &mut Bencher) {
             .clone()
             .iter_mut()
             .filter_map(|b| {
-                *b = TO_DNA_IUPAC_NO_GAPS_UC[*b as usize];
+                *b = TO_DNA_IUPAC_NO_GAPS_UC[*b];
                 if *b > 0 { Some(*b) } else { None }
             })
             .collect();
